fuck off matrix,i hate math Signed-off-by: MiaoMiao <360382992@qq.com>

+++ JGame/RenderSystem/Math/Rect.java
@@ -180,7 +180,7 @@ public class Rect {
 
 		@Override
 		public String toString() {
-			return "Rect [x=" + x + ", y=" + y + ", width=" + width + ", height=" + height + "]";
+			return "Rect [x=" + (int)x + ", y=" + (int)y + ", width=" + (int)width + ", height=" + (int)height + "]";
 		}
